pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "optimal-bidding-model"
version = "0.1.0"
dependencies = [
  "pandas",
  "openpyxl",
  "PyYAML",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]